fix: serialize decimal relation score symmetrically so save round-trip is byte-stable (#3753)
DiplomacyRelation.fromJson migrates legacy integer scores to decimal
(× 1.0) per SPEC, but toJson emitted the raw runtime value, so a
fresh-init integer score (50) and a loaded decimal score (50.0)
serialized differently. This broke the seed-42 game fixture lossless
round-trip guard (App tests shard 0/2).

Emit score.toDouble() in toJson so the persisted form always matches
fromJson, and regenerate the committed seed-42 fixture scores to the
decimal form. SPEC/game/diplomacy.md § Relation Model already specifies
save/load round-trips the decimal (50 loads as 50.0).

Refs #3753

+++ packages/colonizethis_models/lib/src/diplomacy.dart
@@ -72,7 +72,11 @@ class DiplomacyRelation {
   Map<String, dynamic> toJson() => {
     'factionId1': factionId1,
     'factionId2': factionId2,
-    'score': score,
+    // Serialize the decimal form (× 1.0) so save/load round-trips the decimal
+    // symmetrically with `fromJson`: a whole score of 50 persists as `50.0`,
+    // not `50`, keeping load → re-encode byte-stable.
+    // SPEC/game/diplomacy.md § Relation Model.
+    'score': score.toDouble(),
     'level': level.name,
     'state': state.name,
     'sinceTurn': sinceTurn,
